pyproject.toml
[tool.poetry]
name = "streamlit-image-classification"
version = "0.1.0"
description = ""
authors = ["salsifitewn <salsifite@gmail.com>"]
readme = "README.md"
packages = [{include = "streamlit_image_classification"}]

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.39.1"
streamlit = "^1.32.2"
pandas = "^2.2.1"
scikit-learn = "^1.4.1.post1"
torch = "^2.2.1"
deep-translator = "^1.11.4"

[tool.poetry.group.dev.dependencies]
pillow = "^10.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
